+++ libnetdata/libnetdata.h
@@ -836,6 +836,7 @@ extern char *netdata_configured_host_prefix;
 #include "yaml.h"
 #include "http/http_defs.h"
 #include "gorilla/gorilla.h"
+#include "facets/facets.h"
 #include "dyn_conf/dyn_conf.h"
 
 // BEWARE: this exists in alarm-notify.sh
